Inbound fix and Outbound refactor
