+++ packages/invoices/src/InboundShipment/api/hooks/line/useInboundLines.ts
@@ -1,7 +1,22 @@
 import { useCallback } from 'react';
 import { isA } from './../../../../utils';
 import { InboundFragment } from '../../operations.generated';
-import { useInboundSelector } from './index';
+import { useQuery } from 'packages/common/src';
+import { useInboundNumber } from '../document/useInbound';
+import { useInboundApi } from '../utils/useInboundApi';
+
+export const useInboundSelector = <T = InboundFragment>(
+  select?: (data: InboundFragment) => T
+) => {
+  const invoiceNumber = useInboundNumber();
+  const api = useInboundApi();
+
+  return useQuery(
+    api.keys.detail(invoiceNumber),
+    () => api.get.byNumber(invoiceNumber),
+    { select }
+  );
+};
 
 export const useInboundLines = (itemId?: string) => {
   const selectItems = useCallback(
