+++ packages/invoices/src/InboundShipment/api/hooks/line/useInboundServiceLines.ts
@@ -1,7 +1,7 @@
 import { useCallback } from 'react';
 import { isA } from '../../../../utils';
 import { InboundFragment } from '../../operations.generated';
-import { useInboundSelector } from './index';
+import { useInboundSelector } from '../line/useInboundLines';
 
 export const useInboundServiceLines = () => {
   const selectLines = useCallback((invoice: InboundFragment) => {
