+++ packages/invoices/src/OutboundShipment/api/hooks/document/index.ts
@@ -1,8 +1,20 @@
-export * from './useOutbound';
-export * from './useOutboundDelete';
-export * from './useOutboundFields';
-export * from './useOutbounds';
-export * from './useOutboundsAll';
-export * from './useOutboundUpdateTax';
-export * from './useOutboundInsert';
-export * from './useOutboundUpdate';
+import { useOutbound } from './useOutbound';
+import { useOutboundDelete } from './useOutboundDelete';
+import { useOutboundFields } from './useOutboundFields';
+import { useOutbounds } from './useOutbounds';
+import { useOutboundsAll } from './useOutboundsAll';
+import { useUpdateOutboundTax } from './useOutboundUpdateTax';
+('./useOutboundUpdateTax');
+import { useOutboundInsert } from './useOutboundInsert';
+import { useOutboundUpdate } from './useOutboundUpdate';
+
+export const Document = {
+  useOutbound,
+  useOutboundDelete,
+  useOutboundFields,
+  useOutbounds,
+  useOutboundsAll,
+  useUpdateOutboundTax,
+  useOutboundInsert,
+  useOutboundUpdate,
+};
