+++ packages/invoices/src/OutboundShipment/api/hooks/utils/index.ts
@@ -2,9 +2,7 @@ import { useOutboundNumber } from './useOutboundNumber';
 import { useOutboundIsDisabled } from './useOutboundIsDisabled';
 import { useOutboundApi } from './useOutboundApi';
 
-export { useOutboundApi, useOutboundIsDisabled, useOutboundNumber };
-
-export const utils = {
+export const Utils = {
   useOutboundNumber,
   useOutboundIsDisabled,
   useOutboundApi,
